# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  models.Chapter:
    properties:
      name:
        type: string
      num:
        type: string
      paragraphs:
        items:
          $ref: '#/definitions/models.Paragraph'
        type: array
    type: object
  models.Doc:
    properties:
      abbreviation:
        type: string
      chapters:
        items:
          $ref: '#/definitions/models.Chapter'
        type: array
      name:
        type: string
    type: object
  models.Paragraph:
    properties:
      chapter:
        type: string
      editedText:
        items:
          type: string
        type: array
      tables:
        items:
          $ref: '#/definitions/models.ParagraphTable'
        type: array
      text:
        items:
          type: string
        type: array
    type: object
  models.ParagraphTable:
    properties:
      img:
        type: string
      name:
        type: string
      num:
        type: string
    type: object
host: 188.93.210.165:8080
info:
  contact: {}
  description: API Server for Regulations
  title: Regulations API
  version: "1.0"
paths:
  /regulations/{abbreviation}:
    get:
      consumes:
      - application/json
      description: return regulation doc
      operationId: get-doc
      parameters:
      - description: Regulation Abbreviation
        in: query
        name: abbreviation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Doc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: get doc
      tags:
      - getDoc
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"